# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTML client and backend proxy for GREEN-API methods getSettings, getStateInstance, sendMessage and sendFileByUrl.",
        "title": "GREEN-API HTML Client",
        "contact": {
            "name": "Artem Rylskii",
            "url": "https://t.me/Rainz0r",
            "email": "n52rus@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/file": {
            "post": {
                "description": "Sends a file by public URL via GREEN-API. If chatId is sent as a plain phone number, the backend automatically appends @c.us.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Send file by URL",
                "parameters": [
                    {
                        "description": "File payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SendFileByURLRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GREEN-API response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Validation failed or invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to build request or marshal payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to request GREEN-API",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/message": {
            "post": {
                "description": "Sends a text message via GREEN-API. If chatId is sent as a plain phone number, the backend automatically appends @c.us.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Send text message",
                "parameters": [
                    {
                        "description": "Message payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GREEN-API response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Validation failed or invalid request body",
                        "schema": {
                            "$ref": "#/definitions/models.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to build request or marshal payload",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to request GREEN-API",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/settings": {
            "get": {
                "description": "Returns GREEN-API instance settings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Get instance settings",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GREEN-API instance ID",
                        "name": "idInstance",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "GREEN-API instance API token",
                        "name": "apiTokenInstance",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GREEN-API response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing required query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to build request or read response",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to request GREEN-API",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/state": {
            "get": {
                "description": "Returns current GREEN-API instance state.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "green-api"
                ],
                "summary": "Get instance state",
                "parameters": [
                    {
                        "type": "string",
                        "description": "GREEN-API instance ID",
                        "name": "idInstance",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "GREEN-API instance API token",
                        "name": "apiTokenInstance",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "GREEN-API response",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Missing required query parameters",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to build request or read response",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Failed to request GREEN-API",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "idInstance is required"
                }
            }
        },
        "models.SendFileByURLRequest": {
            "type": "object",
            "required": [
                "apiTokenInstance",
                "chatId",
                "fileName",
                "idInstance",
                "urlFile"
            ],
            "properties": {
                "apiTokenInstance": {
                    "description": "Токен доступа к инстансу GREEN-API",
                    "type": "string"
                },
                "caption": {
                    "description": "Необязательная подпись к файлу",
                    "type": "string"
                },
                "chatId": {
                    "description": "Телефон или chatId, например 79991234567 или 79991234567@c.us",
                    "type": "string"
                },
                "fileName": {
                    "description": "Имя файла с расширением, например file.pdf",
                    "type": "string"
                },
                "idInstance": {
                    "description": "Идентификатор инстанса GREEN-API",
                    "type": "string"
                },
                "quotedMessageId": {
                    "description": "ID сообщения, на которое нужно ответить",
                    "type": "string"
                },
                "urlFile": {
                    "description": "Публичная ссылка на файл",
                    "type": "string"
                }
            }
        },
        "models.SendMessageRequest": {
            "type": "object",
            "required": [
                "apiTokenInstance",
                "chatId",
                "idInstance",
                "message"
            ],
            "properties": {
                "apiTokenInstance": {
                    "description": "Токен доступа к инстансу GREEN-API",
                    "type": "string"
                },
                "chatId": {
                    "description": "Телефон или chatId, например 79991234567 или 79991234567@c.us",
                    "type": "string"
                },
                "idInstance": {
                    "description": "Идентификатор инстанса GREEN-API",
                    "type": "string"
                },
                "message": {
                    "description": "Текст сообщения",
                    "type": "string",
                    "maxLength": 20000
                },
                "quotedMessageId": {
                    "description": "ID сообщения, на которое нужно ответить",
                    "type": "string"
                }
            }
        },
        "models.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation failed"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ValidationFieldError"
                    }
                }
            }
        },
        "models.ValidationFieldError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "chatId is required"
                },
                "field": {
                    "type": "string",
                    "example": "chatId"
                },
                "param": {
                    "type": "string",
                    "example": ""
                },
                "tag": {
                    "type": "string",
                    "example": "required"
                }
            }
        }
    }
}